requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.5
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
